requirements.txt
numpy
sklearn
scipy
attrs
